Changing counter values to store uint64_t values instead of ints.  Also handling the case where xclGetSysfsPath does not find a path to a file and returns null string

+++ src/runtime_src/xdp/profile/database/dynamic_event_database.cpp
@@ -173,7 +173,7 @@ namespace xdp {
   }
 
   void VPDynamicDatabase::addPowerSample(uint64_t deviceId, double timestamp,
-					 const std::vector<int>& values)
+					 const std::vector<uint64_t>& values)
   {
     std::lock_guard<std::mutex> lock(dbLock) ;
